+++ src/metaexpert/core/event_type.py
@@ -7,7 +7,7 @@ from types import ModuleType
 from typing import Self
 
 from metaexpert.core._event_handler import EventHandler
-from metaexpert.logger import BoundLogger, get_logger
+from metaexpert.logger import Logger, get_logger
 from metaexpert.websocket import WebSocketClient
 
 
@@ -120,7 +120,7 @@ class EventType(Enum):
     def __new__(cls, value: dict):
         """Create a new enum member with a copy of the provided value."""
         obj = object.__new__(cls)
-        cls.logger: BoundLogger = get_logger("EventType")
+        cls.logger: Logger = get_logger("EventType")
         # Create a copy of the dictionary and ensure mutable lists are independent
         obj._value_ = {
             k: v if not isinstance(v, list) else [] for k, v in value.items()